Update settings production file in order to use a configuration file
The idea is that the production.py file is considered as part of the
code and doesn't have to be modified by the user/admin.

The file has been modifed in order to parse an ini configuration file
stored in /etc/flopedt directory or in any place with the path given to
the production.py via an environment variable.

All the parameters (database path, redis path, secret key) are parsed
from the configuration file.