Application configuration refactoring
All the configuration is now embedeed into the base file because
most of that configutation is common between development and production.

Now, in base.py we define all django settings parsing a configuration
file which is at system level or defined by an environment variable.

Configuration file has been updated in order to define all the
parameters.