feat(system): Import/Export XLS files scheduling refactoring
- XLS files are now stored into an xls subdirectory of the configuration
module. The idea is that these files are considered part of source
code and not as dynamic content. All links have been updated in the
code in order to link with an absolute path to the XLS file using
CONF_XLS_DIR configuration parameter.

- Define a new directory called storage. (Default path:
/var/flopedt/storage). This directory contains all the generated
content from the configuration module (XLS files). A subdir
is generated for that. The uploaded files are stored in the MEDIA_ROOT
directory which is TMP_DIRECTORY.

- Update settings/base.py:
  * Handle new flopedt.ini parameter storage_directory
  * Define MEDIA_ROOT to TMP_DIRECTORY
  * Define CONF_XLS_DIR parameter
  * Update flopedt.ini templates

- Update base/apps.py in order to create new directories

- Update packaging

Fixes: #65