fix(packaging): Launch flopedt in unbufferized mode with systemd
print messages from the solve_board are bufferized and buffer is not
flushed fast enough with FlOp is started with systemd.

Fortunately, there is a unbuffer parameter of the python interpreter in
order to launch the program in unbufferized mode.

That commits update the launch command in the systemd unit in order for
flop to have the correct behavior

Fixes: #70